adding : each user have the right to only update and delete their contact, not other's

+++ 05-contact-manager/controllers/contactController.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler')
 const Contact = require('../models/contactModel')
 
 const getAllContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.userId});
     res.status(200).json(contacts);
 });
 
@@ -17,7 +17,9 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.userId,
     });
+
 
     res.status(200).json(contact);
 });
@@ -36,6 +38,18 @@ const getContact = asyncHandler(async (req, res) => {
 
 
 const updateContact = asyncHandler(async (req, res) => {
+    const check = await Contact.findById(req.params.id);
+    if (!check) {
+        res.status(404);
+        throw new Error("Contact Not Found")
+    }
+
+    if (check.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error('User don`t have permission to update this contact')
+    }
+
+
     const contact = await Contact.findByIdAndUpdate(req.params.id, req.body, {new: true});
     if (!contact) {
         res.status(404);
@@ -47,6 +61,17 @@ const updateContact = asyncHandler(async (req, res) => {
 
 
 const deleteContact = asyncHandler(async (req, res) => {
+    const check = await Contact.findById(req.params.id);
+    if (!check) {
+        res.status(404);
+        throw new Error("Contact Not Found")
+    }
+
+    if (check.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error('User don`t have permission to update this contact')
+    }
+
     const contact = await Contact.findByIdAndDelete(req.params.id);
     if (!contact) {
         res.status(404);
